my_readlink() wraps readlink(2), to avoid having to look at its rather smelly interface elsewhere in the code (now more than one place).
my_realpath() does realpath(3) into alloc.h-managed memory, and has
flags bits for following links and cutting the transformation short if
an error is detected.

alloc.h gains allocmore(), similar in spirit to realloc(3), to avoid
huge amounts of copying in my_readlink and my_realpath.  Unlike
realloc(3) it guarantees 0 fill of the new space.

append/appendn are used by my_realpath() to build up the result
string.


The consequence of all this keybashing is that sftp_readlink() and
sftp_v345_realpath() are simplified somewhat and we're in a better
position do SFTP v6.