configure: disable test for unaligned access support.
The test as it stands assumes that if unaligned accesses are not
supported then the code will fault.  But this is not true; it may
produce the wrong answer (e.g. round the address down) or be emulated
by the kernel (making it pointless as an optimization).

The only reasonable way to do this is to enable it on platforms where
it's specifically known to work.