Tidy up unaligned data access functions.
UNALIGNED_ACCESS is completely gone.  Knowledge is built into the
source code rather than the configure script.

putN are now inline functions and are complemented by getN.  This
allows incorporation of the byte-swapping with the reading and
dodges compilers that think unaligned accesses are a licence to
do something stupid.