Fix and test behavior with empty paths
The SFTP spec requires that empty paths are treated as referring to
the default directory.  For the most part we implement this by
converting "" to "." just after parsing a path.

The exception is that for symlink destinations, we reject empty paths.
Some kind of translation is required, since the OS will reject an
empty target.  Rather than set a policy, we reject the attempt with
a (somewhat) informative diagnostic; the user will have to make their
own decision.

In order to be able to test this change properly, sftpclient no longer
normalizes paths.

Issue was reported by Stef Bon.

closes #3